Refactoring app
- Move storage hanglers to App\Storage
- Move rendering of page to App\Pages\Homepage
