+++ app/Storage/JsonStorage.php
@@ -1,7 +1,7 @@
 <?php
 declare (strict_types = 1);
 
-namespace App;
+namespace App\Storage;
 
 use Exception;
