+++ www/index.php
@@ -1,18 +1,9 @@
 <?php
 declare (strict_types = 1);
 
+use App\Pages\Homepage;
+
 require __DIR__ . '/../vendor/autoload.php';
 
-use App\Helpers;
-
-Helpers::printNiceHtmlHeader();
-
-$items = Helpers::getItems();
-
-$id = Helpers::getId();
-
-if ($id !== null) {
-    Helpers::printValueFromItems($items, $id);
-} else {
-    Helpers::printForm();
-}
+$homepage = new Homepage();
+$homepage->render();
